requirements.txt
pefile==2022.5.30
pycryptodome==3.16.0
python_magic==0.4.27
